use blockchain cf's to record block existence
We now use CF's in blockchain module to record whether a block is
in the system.  To do so, we only write the writebatch at the end
of the insert_block function, so that upcoming blocks will only
be processed after all metadata updates of the previous block have
completed.  This check is used for all three types of blocks, and
should remove a bunch of (very rare) race conditions in the code.